fix some NPOI.XSSF compilation issue;
git-svn-id: http://npoi.googlecode.com/svn/trunk@355 3d0cb27c-5942-9f1f-b0c8-b51bd5f61ba7

+++ ooxml/XSSF/UserModel/XSSFFont.cs
@@ -575,7 +575,7 @@ namespace NPOI.XSSF.UserModel
         public FontScheme GetScheme()
         {
             NPOI.OpenXmlFormats.Spreadsheet.CT_FontScheme scheme = _ctFont.sizeOfSchemeArray() == 0 ? null : _ctFont.GetSchemeArray(0);
-            return scheme == null ? FontScheme.NONE : FontScheme.ValueOf(scheme.val);
+            return scheme == null ? FontScheme.NONE : FontScheme.ValueOf((int)scheme.val);
         }
 
         /**
